stop-ai-agent-hallucinations/02-semantic-tools-demo/requirements.txt
strands-agents
strands-agents[openai]
strands-agents-tools
sentence-transformers
faiss-cpu
neo4j
jupyter
ipykernel
